[RoiStatsWidget] move ROIStatsWidget to silx.gui.plot
