+++ silx/gui/plot/ROIStatsWidget.py
